Moved helping functions to misc/free_help.py
The helping functions recursive_simpson and plot_FE have been moved to misc such that they cannot be called by inft.
